tests: move postgres database creation to a separate class
EPMDJ-10739
